// packages/sdk-next/src/pool/offlinePoolService/types.ts
export enum AssetType {
  Bond = 'Bond',
  External = 'External',
  StableSwap = 'StableSwap',
  Token = 'Token',
  XYK = 'XYK',
  Erc20 = 'Erc20',
  PoolShare = 'PoolShare',
}

export type AssetDynamicFee = {
  assetFee: number;
  protocolFee: number;
  timestamp: number;
};

export interface PersistentAsset {
  id: string;
  decimals: number;
  symbol: string;
  existentialDeposit: string;
  isSufficient: boolean;
  type: AssetType;
  name?: string;
  icon?: string;
  location?: unknown;
  dynamicFee?: AssetDynamicFee;
}
// ...
export type PersistentDynamicFeesAssetFeeParams = {
  minFee: number;
  maxFee: number;
  decay: string;
  amplification: string;
};

export interface IPersistentConstants {
  lbpRepayFee: number[];
  lbpMaxInRatio: string;
  lbpMaxOutRatio: string;
  lbpMinPoolLiquidity: string;
  lbpMinTradingLimit: string;

  omnipoolBurnProtocolFee: number;
  omnipoolHdxAssetId: number;
  omnipoolHubAssetId: number;
  omnipoolMaxInRatio: string;
  omnipoolMaxOutRatio: string;
  omnipoolMinimumPoolLiquidity: string;
  omnipoolMinimumTradingLimit: string;
  omnipoolMinWithdrawalFee: number;

  stableswapMinTradingLimit: string;
  stableswapMinPoolLiquidity: string;
  stableswapAmplificationRange: number[];

  xykGetExchangeFee: number[];
  xykMaxInRatio: string;
  xykMaxOutRatio: string;
  xykMinPoolLiquidity: string;
  xykMinTradingLimit: string;
  xykNativeAssetId: number;
  xykOracleSource: string;

  dynamicFeesAssetFeeParameters: PersistentDynamicFeesAssetFeeParams;
  dynamicFeesProtocolFeeParameters: PersistentDynamicFeesAssetFeeParams;
}

export interface IPersistentOmnipoolExtras {
  maxSlipFee: number;
}

export interface IPersistentExtras {
  omnipool: IPersistentOmnipoolExtras;
}

export type EmaOracleSource = string;
export type EmaOraclePeriod = string;

export interface IPersistentEmaOracleEntry {
  source: EmaOracleSource;
  period: EmaOraclePeriod;
  assets: string[];
  entry: IPersistentEmaOracleEntryData;
}

export interface IPersistentEmaOracleEntryData {
  price: IPersistentEmaOracleEntryRatio;
  volume: IPersistentEmaOracleEntryVolume;
  liquidity: IPersistentEmaOracleEntryLiquidity;
  updatedAt: number;
}

export interface IPersistentEmaOracleEntryLiquidity {
  a: string;
  b: string;
}

export interface IPersistentEmaOracleEntryVolume {
  aIn: string;
  bOut: string;
  aOut: string;
  bIn: string;
}

export interface IPersistentEmaOracleEntryRatio {
  n: string;
  d: string;
}

export interface IPersistentMmOracleEntry {
  address: string;
  price: string;
  decimals: number;
  updatedAt: number;
}
// ...
export interface IOfflinePoolServiceDataSource {
  assets: Array<PersistentAsset>;
  pools: {
    lbp: Array<LbpPoolBase>;
    xyk: Array<PoolBase>;
    stableswap: Array<StableSwapBase>;
    omnipool: Array<OmniPoolBase>;
    aave: Array<PoolBase>;
  };
  extras: IPersistentExtras;
  constants: IPersistentConstants;
  emaOracle: Array<IPersistentEmaOracleEntry>;
  mmOracle: Array<IPersistentMmOracleEntry>;
  meta: IPersistentMetaData;
}

export interface IPersistentMetaData {
  paraBlockNumber: number;
  paraBlockHash: string;
  relayBlockNumber: number;
}